+++ public/index.php
@@ -4,31 +4,10 @@ use Helpers\Router\Router;
 
 const BASE_PATH = __DIR__ . '/../';
 
-(new App\Controller\DotEnvEnvironment)->load(BASE_PATH);
-
+require 'autoload.php';
 require BASE_PATH . 'helpers/functions.php';
 
-spl_autoload_register(static function ($class) {
-    $namespace = explode('\\', $class);
-    $class = strtolower(str_replace(['\\'], ['/'], array_pop($namespace)));
-
-    // Define the directories to search for classes
-    $directories = [
-        'helpers',
-        'routes',
-        'config',
-        'database',
-        'app/enums'
-    ];
-
-    foreach ($directories as $directory) {
-        $path = BASE_PATH . $directory . '/' . $class . '.php';
-        if (file_exists($path)) {
-            require $path;
-            break;
-        }
-    }
-});
+(new App\Controller\DotEnvEnvironment)->load(BASE_PATH);
 
 $uri = trim(
     str_replace(
